had linked with care and like with userId

+++ src/components/Cart.js
@@ -2,8 +2,11 @@
 import { useContext, useEffect, useState } from "react"
 import DisplayItemListCarOrLike from "./DisplayItemListCarOrLike"
 import axios from "axios";
+import { CounterContext } from "../contextAPI/CreateContextAPI";
 
 export default function Cart() {
+
+  const { newUser } = useContext(CounterContext);
 
 
     const [listOfCart, setListOfCart] = useState([])
@@ -21,17 +24,18 @@ export default function Cart() {
     },[])
 
     const deleteItem = (item) => {
-      // Create a new array excluding the item at the specified index
       const updatedList = listOfCart.filter((existingItem) => existingItem !== item);
       setListOfCart(updatedList);
     };
-
-
-    return (
-        <div>
-            {listOfCart.map((item, index) => (
-                <DisplayItemListCarOrLike key={index} item={item} text={'Proceed to checkout'} onDelete={deleteItem} />
-            ))}
-        </div>
-    )
+  
+  return (
+    <div>
+        {listOfCart.map((item, index) => (
+          item.userId !== null && item.userId === newUser.userId ?(
+            <DisplayItemListCarOrLike key={index} item={item} text={'Proceed to checkout'} onDelete={deleteItem}/> ) :
+            // 
+            ( item.userId === undefined && localStorage.getItem('userId') === null? (<DisplayItemListCarOrLike key={index} item={item} text={'Proceed to checkout'} onDelete={deleteItem}/>) : null )
+        ))}
+    </div>
+)
 }
